requirements.txt
sgp4>=2.25
numpy>=1.24
matplotlib>=3.10
PyYAML>=6.0
scipy
cartopy